Added username validation on login first step
