+++ Identity/Pages/Account/Login/ExternalProvider.cs
@@ -1,7 +1,47 @@
-﻿namespace Identity.Pages.Login;
+﻿using Duende.IdentityServer.Models;
+
+using Microsoft.AspNetCore.Authentication;
+
+namespace Identity.Pages.Login;
 
 public sealed class ExternalProvider
 {
     public string DisplayName { get; set; }
     public string AuthenticationScheme { get; set; }
+
+    public static IEnumerable<ExternalProvider> GetExternalProviders(IEnumerable<AuthenticationScheme> schemes, IEnumerable<IdentityProviderName> identityProviderNames)
+    {
+        var providers = GetExternalProviders(schemes).ToList();
+        var identityProviders = GetExternalProviders(identityProviderNames);
+
+        providers.AddRange(identityProviders);
+
+        return providers;
+    }
+
+    private static IEnumerable<ExternalProvider> GetExternalProviders(IEnumerable<AuthenticationScheme> schemes)
+    {
+        var providers = schemes
+            .Where(x => x.DisplayName != null)
+            .Select(x => new ExternalProvider
+            {
+                DisplayName = x.DisplayName ?? x.Name,
+                AuthenticationScheme = x.Name
+            });
+
+        return providers;
+    }
+
+    private static IEnumerable<ExternalProvider> GetExternalProviders(IEnumerable<IdentityProviderName> identityProviderNames)
+    {
+        var providers = identityProviderNames
+            .Where(x => x.Enabled)
+            .Select(x => new ExternalProvider
+            {
+                AuthenticationScheme = x.Scheme,
+                DisplayName = x.DisplayName
+            });
+
+        return providers;
+    }
 }
