+++ Identity/Pages/Account/Login/InputModel.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Duende Software. All rights reserved.
 // See LICENSE in the project root for license information.
 
+using Microsoft.AspNetCore.Mvc;
+
 using System.ComponentModel.DataAnnotations;
 
 namespace Identity.Pages.Login;
@@ -9,17 +11,8 @@ public class InputModel
 {
     [Required]
     [Display(Name = nameof(Username))]
+    [PageRemote(ErrorMessage = "Username doesn`t exist", HttpMethod = "post", PageHandler = "ValidateUsername")]
     public string Username { get; set; }
 
-    [Required]
-    [Display(Name = nameof(Password))]
-    [DataType(DataType.Password)]
-    public string Password { get; set; }
-
-    [Display(Name = nameof(Password))]
-    public bool RememberLogin { get; set; }
-
     public string ReturnUrl { get; set; }
-
-    public string Button { get; set; }
 }
