+++ Identity/Pages/Account/Login/Password/Index.cshtml
@@ -1,0 +1,4 @@
+﻿@page
+@model Identity.Pages.Login.Password.IndexModel
+
+<h1>Ура, мы на втором этапе</h1>
