+++ Identity/Pages/Account/Login/Password/Index.cshtml.cs
@@ -1,0 +1,13 @@
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace Identity.Pages.Login.Password;
+
+[SecurityHeaders]
+[AllowAnonymous]
+public class IndexModel : PageModel
+{
+    public void OnGet()
+    {
+    }
+}
